refactor(subagents): share local agent path cleanup
