+++ src/local-agent-adapters.ts
@@ -1,8 +1,8 @@
 import { spawn, spawnSync, type ChildProcessWithoutNullStreams } from "node:child_process";
-import { existsSync, readFileSync } from "node:fs";
-import { delimiter, resolve, sep } from "node:path";
+import { resolve } from "node:path";
 import { Readable, Writable } from "node:stream";
 import type { LocalAgentProvider } from "./local-agent-profiles.js";
+import { removeDevspaceNodeModulesBinFromPath } from "./local-agent-path.js";
 import {
   createCodexSdkLocalAgentRuntime,
   type LocalAgentRunInput,
@@ -297,28 +297,8 @@ export function piCommandEnvironment(env: NodeJS.ProcessEnv): NodeJS.ProcessEnv 
 
   return {
     ...env,
-    PATH: path
-      .split(delimiter)
-      .filter((entry) => entry && !isDevspaceNodeModulesBin(entry))
-      .join(delimiter),
+    PATH: removeDevspaceNodeModulesBinFromPath(path),
   };
-}
-
-function isDevspaceNodeModulesBin(pathEntry: string): boolean {
-  const resolvedEntry = resolve(pathEntry);
-  if (!resolvedEntry.endsWith(`${sep}node_modules${sep}.bin`)) {
-    return false;
-  }
-
-  const packageJson = resolve(resolvedEntry, "..", "..", "package.json");
-  if (!existsSync(packageJson)) return false;
-
-  try {
-    const packageInfo = JSON.parse(readFileSync(packageJson, "utf8")) as { name?: unknown };
-    return packageInfo.name === "@waishnav/devspace";
-  } catch {
-    return false;
-  }
 }
 
 class JsonLineRpc {
